Fix TODO
