+++ Source/EasyNetQ.Management.Client/Model/BackingQueueStatus.cs
@@ -1,10 +1,13 @@
+using System.Text.Json.Serialization;
+using EasyNetQ.Management.Client.Serialization;
+
 namespace EasyNetQ.Management.Client.Model;
 
 public record BackingQueueStatus
 (
     int Q1,
     int Q2,
-    // TODO Custom converter is needed
+    [property: JsonConverter(typeof(ObjectReadOnlyListConverter))]
     IReadOnlyList<object?> Delta,
     int Q3,
     int Q4,
